==> Old_Code/01_01_cerebral_patients.ipynb <==
"""
Σε αυτό το αρχείο χρεισιμοποιούνται οι ICD9 κωδικοί που αναφέρονται στο εγκεφαλικό και παράγεται το αρχείο "icd9_cerebral_infarction.csv" όπου τους επεξηγεί.
Στη συνέχεια παράγεται το αρχείο "diagnoses_cerebral_infarction.csv" όπου περιλαμβάνονται σε αυτό οι ασθενείς με εγκεφαλικό.
Έπειτα δημιουργείται το αρχείο "death.csv" όπου εμπεριέχονται σε αυτό οι ασθενείς που απεβίωσαν.

523.740 -> Εισαγωγές στις νοσοκομειακές μονάδες.
 12.015 -> Εισαγωγές στις νοσοκομειακές μονάδες ασθενών με εγκεφαλικό.
  9.926 -> Μοναδικές εισαγωγές στις νοσοκομειακές μονάδες χωρίς περιλαμβάνονται οι ενδονοσοκομειακές μετακινήσεις.
  8.320 -> Μοναδικών ασθενών στις νοσοκομειακές μονάδες.
  7.431 -> Επιβίωσαν στις νοσοκομειακές μονάδες (89,31%).
    889 -> Απεβίωσαν στις νοσοκομειακές μονάδες (10,69%).
  
  
  4.299 -> Εισαγωγές στο σύνολο ασθενών στις ΜΕΘ.
  3.809 -> Μοναδικές εισαγωγές ασθενών στις ΜΕΘ
  3.646 -> Μοναδικοί ασθενείς στις ΜΕΘ
  2.887 -> Επιβίωσαν στις ΜΕΘ (73,71%)
    759 -> Απεβίωσαν στις ΜΕΘ (26,29%)
"""

# Εισαγωγή βιβλιοθήκης Panda
import pandas as pd

# Δήλωση των data frame.
df_icd = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\hosp\d_icd_diagnoses.csv.gz')
df_patients = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\hosp\diagnoses_icd.csv.gz')
df_icu_stay = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\icu\icustays.csv.gz', usecols = ['subject_id', 'hadm_id'])

#Μέγιστη μέγεθος αλφαριθμητικών στις στήλες.
pd.options.display.max_colwidth = 130


#Κωδικοί icd-9 εγκεφαλικού επεισοδίου.
cerebral_infarction_icd9 = df_icd [
    (df_icd['icd_code'] =="43300") | (df_icd['icd_code'] =="43301") | (df_icd['icd_code'] =="43310") | (df_icd['icd_code'] =="43311") | (df_icd['icd_code'] =="43320") |
    (df_icd['icd_code'] =="43321") | (df_icd['icd_code'] =="43330") | (df_icd['icd_code'] =="43331") | (df_icd['icd_code'] =="43380") | (df_icd['icd_code'] =="43381") |
    (df_icd['icd_code'] =="43390") | (df_icd['icd_code'] =="43391") | (df_icd['icd_code'] =="43400") | (df_icd['icd_code'] =="43401") | (df_icd['icd_code'] =="43410") |
    (df_icd['icd_code'] =="43411") | (df_icd['icd_code'] =="43491") | (df_icd['icd_code'] =="43490") | (df_icd['icd_code'] =="436") | (df_icd['icd_code'] =="431") |
    (df_icd['icd_code'] =="430") | (df_icd['icd_code'] =="4350") | (df_icd['icd_code'] =="4351") | (df_icd['icd_code'] =="4352") | (df_icd['icd_code'] =="4353") |
    (df_icd['icd_code'] =="4358") | (df_icd['icd_code'] =="4359")
]

#Ασθενείς με εγκεφαλικό.
cerebral_infarction_diagnoses = df_patients[
    (df_patients['icd_code'] =="43300") | (df_patients['icd_code'] =="43301") | (df_patients['icd_code'] =="43310") | (df_patients['icd_code'] =="43311") |
    (df_patients['icd_code'] =="43320") | (df_patients['icd_code'] =="43321") | (df_patients['icd_code'] =="43330") | (df_patients['icd_code'] =="43331") |
    (df_patients['icd_code'] =="43380") | (df_patients['icd_code'] =="43381") | (df_patients['icd_code'] =="43390") | (df_patients['icd_code'] =="43391") |
    (df_patients['icd_code'] =="43400") | (df_patients['icd_code'] =="43401") | (df_patients['icd_code'] =="43410") | (df_patients['icd_code'] =="43411") |
    (df_patients['icd_code'] =="43491") | (df_patients['icd_code'] =="43490") | (df_patients['icd_code'] =="436") | (df_patients['icd_code'] =="431") |
    (df_patients['icd_code'] =="430") | (df_patients['icd_code'] =="4350") | (df_patients['icd_code'] =="4351") | (df_patients['icd_code'] =="4352") |
    (df_patients['icd_code'] =="4353") | (df_patients['icd_code'] =="4358") | (df_patients['icd_code'] =="4359")
]

#Δημιουργία CSV με τους κωδικούς icd-9 για το εγκεφαλικ΄ό.
cerebral_infarction_icd9.to_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\o1_icd9_brain_cerebral_infarction.csv', index=False)

#Δημιουργία CSV ασθενών με εγκεφαλικό απαλαγμένο από διπλές ενδοσοκομειακές εγγραφές λόγω αλλαγής seq_num
cerebral_infarction_diagnoses.drop_duplicates(subset=['hadm_id']).to_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\o2_diagnoses_cerebral_infarction.csv', index=False)

#Σύνολο εισαγωγών στο νοσοκομείο.
df_admissions = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\core\Admissions.csv.gz', usecols = ['subject_id', 'hadm_id','hospital_expire_flag'])

# Μοναδικές διαγνώσεις εγκεφαλικού.
df_cerebral_infarction_diagnoses = cerebral_infarction_diagnoses[["subject_id", "hadm_id"]].drop_duplicates(subset=['hadm_id'])

# Σύνδεση εισαγωγών ασθενών με εγκεφαλικό με σημαία θνησιμότητας.
df_cerebral_infarction_admissions=pd.merge(df_cerebral_infarction_diagnoses, df_admissions, on=['subject_id','hadm_id'])

deaths = df_cerebral_infarction_admissions[(df_cerebral_infarction_admissions['hospital_expire_flag'] == 1)]

#Δημιουργία dataframe απαλλαγμένο από διπλοεγγραφές λόγω seq_num ασθενών που απεβίωσαν.
final_deaths=deaths.drop_duplicates(subset=['subject_id'])

#Δημιουργία CSV ασθενών που εισήχθησαν με εγκεφαλικό στα νοσοκομεία και απεβίωσαν.
final_deaths.to_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\o3_hospital_deaths.csv', index=False)

# Συνολικές εισαγωγές στις ΜΕΘ (4299)
rslt_df=pd.merge(df_cerebral_infarction_diagnoses, df_icu_stay, on=['subject_id','hadm_id'])

# Σύνδεση εισαγωγών στις ΜΕΘ ασθενών με εγκεφαλικό με σημαία θνησιμότητας.
df_icu_cerebral_infarction_admissions=pd.merge(rslt_df, df_admissions, on=['subject_id','hadm_id'])

icu_deaths = df_icu_cerebral_infarction_admissions[(df_icu_cerebral_infarction_admissions['hospital_expire_flag'] == 1)]

#Δημιουργία dataframe απαλλαγμένο από διπλοεγγραφές λόγω seq_num ασθενών που απεβίωσαν.
final_icu_deaths=icu_deaths.drop_duplicates(subset=['subject_id'])

#Δημιουργία CSV ασθενών με εγκεφαλικό στις ΜΕΘ που απεβίωσαν
final_deaths.to_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\o4_icu_deaths.csv', index=False)

# Μοναδικ΄ές εισαγωγές στις ΜΕΘ (3809)
rslt_df.drop_duplicates(subset=['hadm_id']).to_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\o5_icu_unique_admissions.csv', index=False)

# Μοναδικοί ασθενείς στις ΜΕΘ (3646)
rslt_df.drop_duplicates(subset=['subject_id']).to_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\o6_icu_unique_patients.csv', index=False)
