==> Old_Code/02_01_cerebral_chartevents.ipynb <==
"""
Έχοντας τμηματοποιήσει έξω από την python τον πίνακα chartevents σε επτά διαφορετικούς πίνακες τους συγκρίνω και ενσωματώνω τα στοιχεία τους
στο "diagnoses_cerebral_infarction.csv" όπου εμπεριέχονται οι ασθενείς με εγκεφαλικό που έχουν εισαχθείς στης ΜΕΘ.
"""

# Εισαγωγή βιβλιοθήκης Panda
import pandas as pd

df_patients = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\o5_icu_unique_admissions.csv')

df_patients

#Επεξεργασία πρώτου μέρους του αρχείου chartevent
 
df_chartevents = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\Splited_Chartevent\chartevents_0.csv')

rslt_df=pd.merge(df_patients, df_chartevents, on=['subject_id','hadm_id'])

rslt_df.to_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\Merged_Chartevent\cerebral_infarction_chartevent_0.csv', index=False)

del (df_chartevents, rslt_df) # Διαγραφή των dataframe για απελευθέρωση μνήμης.

#Επεξεργασία δεύτερου μέρους του αρχείου chartevent

df_chartevents = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\Splited_Chartevent\chartevents_1.csv',
                             names = ['subject_id','hadm_id','seq_num','icd_code','icd_version','stay_id','charttime','storetime','itemid','value','valuenum','valueuom','warning']
                            )

rslt_df=pd.merge(df_patients, df_chartevents, on=['subject_id','hadm_id'])

rslt_df.to_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\Merged_Chartevent\cerebral_infarction_chartevent_1.csv', index=False)

del (df_chartevents, rslt_df) # Διαγραφή των dataframe για απελευθέρωση μνήμης.

#Επεξεργασία τρίτου μέρους του αρχείου chartevent

df_chartevents = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\Splited_Chartevent\chartevents_2.csv',
                             names = ['subject_id','hadm_id','seq_num','icd_code','icd_version','stay_id','charttime','storetime','itemid','value','valuenum','valueuom','warning']
                            )

rslt_df=pd.merge(df_patients, df_chartevents, on=['subject_id','hadm_id'])

rslt_df.to_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\Merged_Chartevent\cerebral_infarction_chartevent_2.csv', index=False)

del (df_chartevents, rslt_df) # Διαγραφή των dataframe για απελευθέρωση μνήμης.

#Επεξεργασία τέταρτου μέρους του αρχείου chartevent

df_chartevents = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\Splited_Chartevent\chartevents_3.csv',
                             names = ['subject_id','hadm_id','seq_num','icd_code','icd_version','stay_id','charttime','storetime','itemid','value','valuenum','valueuom','warning']
                            )

rslt_df=pd.merge(df_patients, df_chartevents, on=['subject_id','hadm_id'])

rslt_df.to_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\Merged_Chartevent\cerebral_infarction_chartevent_3.csv', index=False)

del (df_chartevents, rslt_df) # Διαγραφή των dataframe για απελευθέρωση μνήμης.

#Επεξεργασία πέμπτου μέρους του αρχείου chartevent

df_chartevents = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\Splited_Chartevent\chartevents_4.csv',
                             names = ['subject_id','hadm_id','seq_num','icd_code','icd_version','stay_id','charttime','storetime','itemid','value','valuenum','valueuom','warning']
                            )

rslt_df=pd.merge(df_patients, df_chartevents, on=['subject_id','hadm_id'])

rslt_df.to_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\Merged_Chartevent\cerebral_infarction_chartevent_4.csv', index=False)

del (df_chartevents, rslt_df) # Διαγραφή των dataframe για απελευθέρωση μνήμης.

#Επεξεργασία έκτου μέρους του αρχείου chartevent

df_chartevents = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\Splited_Chartevent\chartevents_5.csv',
                             names = ['subject_id','hadm_id','seq_num','icd_code','icd_version','stay_id','charttime','storetime','itemid','value','valuenum','valueuom','warning']
                            )

rslt_df=pd.merge(df_patients, df_chartevents, on=['subject_id','hadm_id'])

rslt_df.to_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\Merged_Chartevent\cerebral_infarction_chartevent_5.csv', index=False)

del (df_chartevents, rslt_df) # Διαγραφή των dataframe για απελευθέρωση μνήμης.

#Επεξεργασία έβδομου μέρους του αρχείου chartevent

df_chartevents = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\Splited_Chartevent\chartevents_6.csv',
                             names = ['subject_id','hadm_id','seq_num','icd_code','icd_version','stay_id','charttime','storetime','itemid','value','valuenum','valueuom','warning']
                            )

rslt_df=pd.merge(df_patients, df_chartevents, on=['subject_id','hadm_id'])

rslt_df.to_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\Merged_Chartevent\cerebral_infarction_chartevent_6.csv', index=False)

del (df_chartevents, rslt_df) # Διαγραφή των dataframe για απελευθέρωση μνήμης.

"""
ΠΡΟΣΟΧΗ ΣΕ ΑΥΤΟ ΤΟ ΣΗΜΕΙΟ ΠΡΕΠΕΙ ΝΑ ΠΑΡΑΜΕΤΡΟΠΟΙΗΣΟΥΜΕ ΤΙΣ ΕΠΙΚΕΦΑΛΙΔΕΣ ΤΩΝ ΠΑΡΑΓΩΜΕΝΩΝ ΑΡΧΕΙΩΝ!!!
Ανοίγω το πρώτο csv που έχει δημιουργηθεί με την ονομασία "cerebral_infarction_chartevent_0.csv"
και αντιγράφω την πρώτη γραμμή που περιέχει τους τίτλους των στηλών συνοδευόμενη από την ακολουθεία ,del,del,del
Είναι τυχαία ονόματα που έδωσα στα τρία τελευταία στοιχεία της λίστας μιας και κατά την επεξεργασία αυτών παράχθησαν τρεις επιπλέον στήλες
τις οποίες όμως δεν θα χρησιμοποιήσω.

Στη συνέχεια έχω παραθέσει τον τίτλο που πρέπει να φέρουν όλα τα αρχεία με ονομασία cerebral_infarction_chartevent_X.csv

subject_id,hadm_id,stay_id,charttime,storetime,itemid,value,valuenum,valueuom,warning,del,del,del
"""

#Φορτώνουμε τα επτά αρχεία

df_cerebral_infarction_chartevent_0 = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\Merged_Chartevent\cerebral_infarction_chartevent_0.csv',
                              usecols = ['subject_id','hadm_id','stay_id','charttime','storetime','itemid','value','valuenum','valueuom','warning'])

df_cerebral_infarction_chartevent_1 = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\Merged_Chartevent\cerebral_infarction_chartevent_1.csv',
                              usecols = ['subject_id','hadm_id','stay_id','charttime','storetime','itemid','value','valuenum','valueuom','warning'])

df_cerebral_infarction_chartevent_2 = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\Merged_Chartevent\cerebral_infarction_chartevent_2.csv',
                              usecols = ['subject_id','hadm_id','stay_id','charttime','storetime','itemid','value','valuenum','valueuom','warning'])

df_cerebral_infarction_chartevent_3 = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\Merged_Chartevent\cerebral_infarction_chartevent_3.csv',
                              usecols = ['subject_id','hadm_id','stay_id','charttime','storetime','itemid','value','valuenum','valueuom','warning'])

df_cerebral_infarction_chartevent_4 = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\Merged_Chartevent\cerebral_infarction_chartevent_4.csv',
                              usecols = ['subject_id','hadm_id','stay_id','charttime','storetime','itemid','value','valuenum','valueuom','warning'])

df_cerebral_infarction_chartevent_5 = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\Merged_Chartevent\cerebral_infarction_chartevent_5.csv',
                              usecols = ['subject_id','hadm_id','stay_id','charttime','storetime','itemid','value','valuenum','valueuom','warning'])

df_cerebral_infarction_chartevent_6 = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\Merged_Chartevent\cerebral_infarction_chartevent_6.csv',
                              usecols = ['subject_id','hadm_id','stay_id','charttime','storetime','itemid','value','valuenum','valueuom','warning'])

#Ενσωματώνουμε τα μέρη του chartevents σε ένα αρχείο.
pd.concat(
    [df_cerebral_infarction_chartevent_0,
     df_cerebral_infarction_chartevent_1,
     df_cerebral_infarction_chartevent_2,
     df_cerebral_infarction_chartevent_3,
     df_cerebral_infarction_chartevent_4,
     df_cerebral_infarction_chartevent_5,
     df_cerebral_infarction_chartevent_6]).to_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\o7_cerebral_infarction_chartevent.csv', index=False)

# Διαγραφή των dataframe για απελευθέρωση μνήμης.
del (df_cerebral_infarction_chartevent_0,
     df_cerebral_infarction_chartevent_1,
     df_cerebral_infarction_chartevent_2,
     df_cerebral_infarction_chartevent_3,
     df_cerebral_infarction_chartevent_4,
     df_cerebral_infarction_chartevent_5,
     df_cerebral_infarction_chartevent_6)

# Επικαλούμε και ενοποιώ τον πίνακα μου με τα στοιχεία του πίνακα που επεξηγεί σε τι είδους ζωτικής ένδειξης αναφέρεται το itemid.
# Παρακάτω κάνω εκ νέου αυτή την ενοποίηση οπότε δεν φαίνεται κάπου για την ώρα η χρηστική της σημασία.
df_cerebral_infarction_chartevent = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\o7_cerebral_infarction_chartevent.csv')

df_d_items = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\icu\d_items.csv.gz', usecols = ['itemid', 'label', 'abbreviation', 'category','param_type'])

rslt_df=pd.merge(df_cerebral_infarction_chartevent, df_d_items, on=['itemid'])

rslt_df.to_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\o8_cerebral_infarction_chartevent_item_merged.csv', index=False)

del (rslt_df)
