==> Old_Code/02_02_seperate_nan_rows.ipynb <==
"""
Πρέπει να απομονώσω τις γραμμές με NaN από την στήλη valuenum
μιας και χρειάζομαι μόνο τις αριθμητικές καταγραφές και όχι τα κείμενα.

26.600.037 συνολικά οι εγγραφές για ασθενών στις ΜΕΘ

9.874.474 οι εγγραφές που περιέχουν τιμές
"""

# Εισαγωγή βιβλιοθήκης Panda
import pandas as pd

df_cerebral_infarction_chartevent = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\o8_cerebral_infarction_chartevent_item_merged.csv')

rslt_df = df_cerebral_infarction_chartevent[df_cerebral_infarction_chartevent['valuenum'].notna()]

rslt_df.to_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\o8_cerebral_infarction_chartevent_item_merged_without_nan.csv', index=False)
