==> Old_Code/02_03_chart_used_items.ipynb <==
"""
Κάνουμε καταμέτρηση των αναγνωριστικών κωδικών που εμφανίζονται στο chartevents.

707 Αναγνωριστικοί κωδικοί μέτρησης

* Πριν διαχωρίσω τον πίνακα οι εγγραφές ήταν 1784 και αρκετά κελιά παρέμεναν NaN λόγο των κειμένων που περιείχαν
"""

#Εισαγωγή βιβλιοθήκης Panda
import pandas as pd

df_cerebral_infarction_chartevents = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\o8_cerebral_infarction_chartevent_item_merged_without_nan.csv')
df_data_chartevents = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\icu\d_items.csv.gz')

# Καταμέτρηση των εξετάσεων που κάνουν οι ασθενείς με εγκεφαλικό και μετατροπή από series σε dataframe.
rslt_df = pd.DataFrame(df_cerebral_infarction_chartevents['itemid'].value_counts(dropna=True))

# Το παραγώμενο αρχείο έχει σαν index την στήλη που περιέχει τον κωδικό των εξετάσεων, οπότε την κάνουμε reset
rslt_df = rslt_df.reset_index()

# Και ορίζουμε νέες ονομασίες στις στήλες για να μπορούν να είναι ευανάγνωστες.
rslt_df.rename(columns={'index': 'itemid', 'itemid': 'times'}).to_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\o9_cerebral_infarction_chartevents_counts.csv', index=False)

df_counts = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\o9_cerebral_infarction_chartevents_counts.csv',
                        index_col=False)

df_count_merge = pd.merge(df_data_chartevents, df_counts, on=['itemid'])

#Εξαγωγή σε csv
df_count_merge.to_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\o9_cerebral_infarction_chartevents_counts_with_labels.csv', index=False)

rslt_df = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\o9_cerebral_infarction_chartevents_counts_with_labels.csv',
                      usecols = ['itemid'], index_col=False)

rslt_df = rslt_df.set_index('itemid')

# Μεταφέρω τις γραμμές σε στήλες
rslt_df = rslt_df.transpose()

rslt_df
