==> Old_Code/02_04_mean_cerebral_chartevents.ipynb <==
"""
Το chartevent αποτελείται από 9.874.474 εγγραφές που αφορούν τους ασθενείς με εγκεφαλικό

Οι καταγραφές αυτές ομαδοποιήθηκαν και υπολογίσαμε τον μέσο όρο για την κάθε μια από αυτές μέσα στο 8ώρο όπου και ανήλθαν στα 4.177.458 εγγραφές 
"""

import pandas as pd

rslt_df = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\o8_cerebral_infarction_chartevent_item_merged_without_nan.csv')

df_d_items = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\icu\d_items.csv.gz')

rslt_df

"""
Μετατροπή της στήλης charttime από object σε datetime
Αυτή η γραμμή κώδικα έγινε σε αντικατάσταση της
dates =  pd.to_datetime(df.charttime, format='%Y-%m-%d %H:%M:%S').
Η εντολή .info() στο dataframe και στο δικό μου είναι rslt_df.info()
γυρίζει αποτελέσματα αναφορικά με τα στοιχεία του πίνακα και ένα από
αυτά είναι το είδος της κάθε στήλης. Η στήλη του charttime φαίνεται
σαν object και αυτό είχε σαν συνέπεια να μην μπορούν να γίνουν οι
υπολογισμοί. Μετά την εκτέλεση της παρακάτω εντολής αν τρέξουμε εκ νέου την
.info() πλέον το αποτέλεσμα στην charttime είναι datetime64[ns], οπότε μετά
μπορεί να ακουθήσει η groupby και οι υπολογισμοί μέσα σε αυτή όπως το freq και
οι υπόλοιπες. Χωρίς αυτή δεν γίνεται.
"""
rslt_df['charttime']= pd.to_datetime(rslt_df['charttime'])

# Ομαδοποίηση στοιχείων ανά οχτάωρο
grouped_multiple = rslt_df.groupby(['subject_id', 'hadm_id', 'itemid', pd.Grouper(key='charttime', freq='8H')]).agg({'valuenum': ['mean']})
grouped_multiple.columns = ['Mean_Chart']
grouped_multiple = grouped_multiple.reset_index()

grouped_multiple

# Ενοποίηση του πίνακα με την επεξήγηση του κάθε itemid
rslt_df=pd.merge(grouped_multiple, df_d_items, on=['itemid'])

rslt_df.to_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\o9_charttime_8_hour_mean_item_merged.csv', index=False)
