==> Old_Code/02_06_fill_chart_dataframe_itemid_header.ipynb <==
"""
Γεμίζω τον πίνακα με τις μετρήσεις
του chartevet 8ώρου.

Ο πίνακας μου απαρτίζεται από 707 στήλες και 65189 γραμμές,
οπότε μου δημιουργούν 46.088.623 κελιά.

Οι τιμές μου ανέρχονται στα 4.177.458 οπότε με μια αφαίρεση
προκείπτουν 41.911.165 NaN κελιά
"""

import pandas as pd

df_chart_tbl = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\o9_empty_chart_tbl.csv', index_col=False)
df_data_chartevents = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\icu\d_items.csv.gz')
df_grouped_cerebral_infarction = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\o9_cerebral_infarction_chartevents_grouped_8hours_date_time.csv', index_col=False)

# Φέρνω τις τελευταίες τέσσερις στήλες πρώτες
cols = list(df_chart_tbl.columns)
cols = [cols[-1]] + cols[:-1]
df_chart_tbl = df_chart_tbl[cols]

cols = list(df_chart_tbl.columns)
cols = [cols[-1]] + cols[:-1]
df_chart_tbl = df_chart_tbl[cols]

cols = list(df_chart_tbl.columns)
cols = [cols[-1]] + cols[:-1]
df_chart_tbl = df_chart_tbl[cols]

cols = list(df_chart_tbl.columns)
cols = [cols[-1]] + cols[:-1]
df_chart_tbl = df_chart_tbl[cols]

# Ελέγχω αν υπάρχει NaN στην στήλη με τις μέσες τιμές, το αποτέλεσμα ισούται με μηδέν.
df_grouped_cerebral_infarction['Mean_Chart'].isna().sum()

df_grouped_cerebral_infarction = pd.merge(df_grouped_cerebral_infarction, df_data_chartevents, on=['itemid'])

# Δημιουργώ εκ νέου την στήλη concat χωρίς κενά ανάμεσα στα στοιχεία.
# Θα χρησιμοποιήσω αυτή τη στήλη σαν index γιατί τα κενά μπορεί να δημιουργήσουν πρόβλημα.
df_chart_tbl["concat"] = (df_chart_tbl["subject_id"].astype(str)
                          + df_chart_tbl["hadm_id"].astype(str)
                          + df_chart_tbl["date"].astype(str)
                          + df_chart_tbl["time"].astype(str))

# Δημιουργώ και την στήλη department έτσι ώστε να συμπληρωθεί το τμήμα
# από το οποίο έγινε η καταγραφή. Δεν έχει λειτουργική σημασία στο
# σημείο που είμαστε αλλά θα χρειαστεί ΄στον ενοποιημένο πίνακα.
df_chart_tbl.insert(2, 'Department', '')

# Αφαιρώ τους χαρακτήρες - και : από της τιμές της concat για να αποφύγω πιθανά προβλήματα
df_chart_tbl['concat'] = df_chart_tbl['concat'].str.replace('-','')
df_chart_tbl['concat'] = df_chart_tbl['concat'].str.replace(':','')

# Ορίζω την στήλη concat ως index των γραμμών
df_chart_tbl = df_chart_tbl.set_index('concat')

# Δημιουργώ την στήλη concat και στον πίνακα που περιλαμβάνει τις μετρήσεις για να είναι το κοινό σημείο ανάμεσα στους δυο πίνακες
df_grouped_cerebral_infarction["concat"] = (df_grouped_cerebral_infarction["subject_id"].astype(str)
                                            + df_grouped_cerebral_infarction["hadm_id"].astype(str)
                                            + df_grouped_cerebral_infarction["date"].astype(str)
                                            + df_grouped_cerebral_infarction["time"].astype(str))

#Δημιουργώ και την στήλη Department
df_grouped_cerebral_infarction.insert(0, 'Department', 'icu')

# Αφαιρώ τους χαρακτήρες - και : από της τιμές της concat για να αποφύγω πιθανά προβλήματα
df_grouped_cerebral_infarction['concat'] = df_grouped_cerebral_infarction['concat'].str.replace('-','')
df_grouped_cerebral_infarction['concat'] = df_grouped_cerebral_infarction['concat'].str.replace(':','')

# Μεταφέρω την τελευταία στήλη πρώτη
cols = list(df_grouped_cerebral_infarction.columns)
cols = [cols[-1]] + cols[:-1]
df_grouped_cerebral_infarction = df_grouped_cerebral_infarction[cols]

for row in df_grouped_cerebral_infarction.itertuples():
    con = row.concat
    item = row.itemid
    value = row.Mean_Chart
    df_chart_tbl.at['{}'.format(con), 'Department'] = row.Department
    #print('{}'.format(con), '{}'.format(item), '{}'.format(value))
    df_chart_tbl.at['{}'.format(con), '{}'.format(item)] = value

df_chart_tbl

"""
Ο πίνακας μου απαρτίζεται από 707 στήλες και 65189 γραμμές οπότε μου δημιουργούν 46.088.623 κελιά.

Οι τιμές μου ανέρχονται στα 4.177.458 οπότε με μια αφαίρεση προκείπτουν 41.911.165 NaN κελιά

Όταν οι κεφαλίδες ήταν οι κωδικοί των παρατηρήσεων οι καταγραφές ήταν ολοκληρωμένες,
όταν τις άλλαξε με τις ονομασίες 12 καταγραφές δεν έχουν περαστεί.
"""

# Η εντολή επαληθεύει το παραπάνω αποτέλεσμα.
df_chart_tbl.isna().sum().sum()

# Αποθήκευση του τελικού αρχείου με επικεφαλίδες τους κωδικούς των παρατηρήσεων
df_chart_tbl.to_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\o9_fill_chart_itemid_header_tbl.csv', index=False)
