==> Old_Code/02_07_fill_chart_dataframe_label_header.ipynb <==
"""
Αντικατάσταση των κεφαλίδων στο αρχείο με τις συνολικές
καταγραφές
"""

#Εισαγωγή βιβλιοθήκης Panda
import pandas as pd

df_data_chartevents = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\icu\d_items.csv.gz')
df_item_header = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\o9_fill_chart_itemid_header_tbl.csv')

df_data_chartevents = df_data_chartevents.set_index('itemid')

# Ενοποίηση τιμών σε κελιά
df_data_chartevents["header"] = (df_data_chartevents["label"].astype(str)
                                                + " ("
                                                + df_data_chartevents["unitname"].astype(str)
                                                + ")")

# Ελέγχω τις κεφαλίδες και αν δεν ανήκουν στις
# τέσσερις πρώτες τιμές κάνω τις αντικατάσταση.
for col in df_item_header.columns:
    if ((col != 'subject_id') and (col != 'hadm_id') and (col != 'Department') and (col != 'date') and (col != 'time')):
        temp = (df_data_chartevents.loc[int(col), 'header'])
        df_item_header.rename(columns={'{}'.format(col): '{}'.format(temp)}, inplace=True)

df_item_header

# Αποθήκευση του τελικού αρχείου με επικεφαλίδες τους τίτλους των παρατηρήσεων
df_item_header.to_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\o9_fill_chart_label_header_tbl.csv', index=False)
