==> Old_Code/03_set_death_flag.ipynb <==
"""
Δεν λειτουργεί ακόμα αυτό το αρχείο
"""

import pandas as pd

df_grouped_cerebral_infarction = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\o13_cerebral_infarction_labevents_grouped_8hours_date_time.csv', index_col=False)

df_lab_tbl = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\o14_lab_tbl.csv', index_col=False)

df_lab_tbl

df_admissions = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\core\Admissions.csv.gz', usecols = ['subject_id', 'hadm_id', 'dischtime', 'hospital_expire_flag'])

df_admissions

df_admissions['dischtime']= pd.to_datetime(df_admissions['dischtime'])

#Διαχωρισμός σε δυο καινούργιες στήλες της ημερομηνίας και ώρας
df_admissions['date'] = pd.to_datetime(df_admissions['dischtime']).dt.date
df_admissions['time'] = pd.to_datetime(df_admissions['dischtime']).dt.time

df_admissions = df_admissions[['subject_id', 'hadm_id', 'dischtime', 'date', 'time', 'hospital_expire_flag']]

df_admissions
