==> Old_Code/03_01_merge_mean _labevents.ipynb <==
"""
Απομονώνω από το αρχείο labevents.csv όπου περιλαμβάνει
τα εργαστηριακά αποτελέσματα τους ασθενείς με εγκεφαλικό

Έχοντας τμηματοποιήσει έξω από την python τον πίνακα labevents
σε τρεις διαφορετικούς πίνακες τους συγκρίνω και ενσωματώνω τα
στοιχεία τους στο "diagnoses_cerebral_infarction.csv" όπου εμπεριέχονται
οι ασθενείς με εγκεφαλικό.


Στο σύνολο τους οι εγγραφές της labevent είναι 122.289.828
Από αυτές το 1.568.922 αναφέρονται στους ασθενείς με εγκεφαλικό που
έχουν εισαχθεί στις ΜΕΘ.
Οι 1.410.362 είναι εγγραφές που περιλαμβάνουν τιμές και όχι κείμενα.
Οι 1.276.856 είναι ομαδοποιημένες οι ίδιες εξετάσεις ανά οχτάωρο.
"""

#Εισαγωγή βιβλιοθήκης Panda
import pandas as pd

df_patients = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\o5_icu_unique_admissions.csv')

#Επεξεργασία πρώτου μέρους του αρχείου labevents
 
df_labevents = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\Splited_Labevents\labevents_0.csv')
df_labevents.columns = ["labevent_id", "subject_id", "hadm_id", "specimen_id", "itemid", "charttime", "storetime", "value",
                        "valuenum", "valueuom", "ref_range_lower", "ref_range_upper", "flag", "priority", "comments"]

rslt_df=pd.merge(df_patients, df_labevents, on=['subject_id','hadm_id'])

rslt_df.to_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\Merged_Labevents\cerebral_infarction_labevents_0.csv', index=False)

del (df_labevents, rslt_df) # Διαγραφή των dataframe για απελευθέρωση μνήμης.

#Επεξεργασία δεύτερου μέρους του αρχείου labevents

df_labevents = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\Splited_Labevents\labevents_1.csv')
df_labevents.columns = ["labevent_id", "subject_id", "hadm_id", "specimen_id", "itemid", "charttime", "storetime", "value",
                        "valuenum", "valueuom", "ref_range_lower", "ref_range_upper", "flag", "priority", "comments"]

rslt_df=pd.merge(df_patients, df_labevents, on=['subject_id','hadm_id'])

rslt_df.to_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\Merged_Labevents\cerebral_infarction_labevents_1.csv', index=False)

del (df_labevents, rslt_df) # Διαγραφή των dataframe για απελευθέρωση μνήμης.

#Επεξεργασία τρίτου μέρους του αρχείου labevents

df_labevents = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\Splited_Labevents\labevents_2.csv')
df_labevents.columns = ["labevent_id", "subject_id", "hadm_id", "specimen_id", "itemid", "charttime", "storetime", "value",
                        "valuenum", "valueuom", "ref_range_lower", "ref_range_upper", "flag", "priority", "comments"]

rslt_df=pd.merge(df_patients, df_labevents, on=['subject_id','hadm_id'])

rslt_df.to_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\Merged_Labevents\cerebral_infarction_labevents_2.csv', index=False)

del (df_labevents, rslt_df) # Διαγραφή των dataframe για απελευθέρωση μνήμης.

df_cerebral_infarction_labevents_0 = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\Merged_Labevents\cerebral_infarction_labevents_0.csv',
                                                 usecols = ["labevent_id", "subject_id", "hadm_id", "specimen_id", "itemid", "charttime", "storetime", "value",
                                                            "valuenum", "valueuom", "ref_range_lower", "ref_range_upper", "flag", "priority", "comments"])

df_cerebral_infarction_labevents_1 = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\Merged_Labevents\cerebral_infarction_labevents_1.csv',
                                                 usecols = ["labevent_id", "subject_id", "hadm_id", "specimen_id", "itemid", "charttime", "storetime", "value",
                                                            "valuenum", "valueuom", "ref_range_lower", "ref_range_upper", "flag", "priority", "comments"])

df_cerebral_infarction_labevents_2 = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\Merged_Labevents\cerebral_infarction_labevents_2.csv',
                                                 usecols = ["labevent_id", "subject_id", "hadm_id", "specimen_id", "itemid", "charttime", "storetime", "value",
                                                            "valuenum", "valueuom", "ref_range_lower", "ref_range_upper", "flag", "priority", "comments"])

#Ενσωματώνουμε τα μέρη του chartevents σε ένα αρχείο.
pd.concat(
    [df_cerebral_infarction_labevents_0,
     df_cerebral_infarction_labevents_1,
     df_cerebral_infarction_labevents_2]).to_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\o10_cerebral_infarction_labevents.csv', index=False)

# Διαγραφή των dataframe για απελευθέρωση μνήμης.
del (df_cerebral_infarction_labevents_0,
     df_cerebral_infarction_labevents_1,
     df_cerebral_infarction_labevents_2)

df_cerebral_infarction_labevents = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\o10_cerebral_infarction_labevents.csv')

# Μετατροπή της στήλης charttime από object σε datetime

df_cerebral_infarction_labevents['charttime']= pd.to_datetime(df_cerebral_infarction_labevents['charttime'])

#Απομονώνω τις γραμμές που δεν περιλαμβάνουν τιμές στην valuenum
df_cerebral_infarction_labevents = df_cerebral_infarction_labevents[df_cerebral_infarction_labevents['valuenum'].notna()]

grouped_multiple = df_cerebral_infarction_labevents.groupby(['subject_id', 'hadm_id', 'itemid', pd.Grouper(key='charttime', freq='8H')]).agg({'valuenum': ['mean']})
grouped_multiple.columns = ['Mean_Lab']
grouped_multiple = grouped_multiple.reset_index()
grouped_multiple

grouped_multiple.to_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\o11_cerebral_infarction_labevents_grouped_8hours.csv', index=False)
