==> Old_Code/03_02_lab_used_items.ipynb <==
"""
Κάνουμε καταμέτρηση των πιο συχνά εμφανιζόμενων εξετάσεων.

Στο σύνολο τους 388 εγγραφές
"""

#Εισαγωγή βιβλιοθήκης Panda.
import pandas as pd

# Δήλωση των dataframes

df_cerebral_infarction_labevents = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\o11_cerebral_infarction_labevents_grouped_8hours.csv')
df_data_lab_events = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\hosp\d_labitems.csv.gz')
df_patients = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\o5_icu_unique_admissions.csv', usecols = ['subject_id'])

# Καταμέτρηση των εξετάσεων που κάνουν οι ασθενείς με εγκεφαλικό και μετατροπή από series σε dataframe.
rslt_df = pd.DataFrame(df_cerebral_infarction_labevents['itemid'].value_counts(dropna=True))

# Το παραγώμενο αρχείο έχει σαν index την στήλη που περιέχει τον κωδικό των εξετάσεων, οπότε την κάνουμε reset
rslt_df = rslt_df.reset_index()

# Και ορίζουμε νέες ονομασίες στις στήλες για να μπορούν να είναι ευανάγνωστες.
rslt_df.rename(columns={'index': 'itemid', 'itemid': 'times'}).to_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\o12_cerebral_infarction_labevents_counts.csv', index=False)

rslt_df = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\o12_cerebral_infarction_labevents_counts.csv', index_col=False)

rslt_df = rslt_df.set_index('itemid')

# Μεταφέρω τις γραμμές σε στήλες
rslt_df = rslt_df.transpose()

# Αδειάζω την μία και μοναδική γραμμή του πίνακα
rslt_df = rslt_df.drop(['times'])

rslt_df
