==> Old_Code/03_03 lab_create_table.ipynb <==
"""
Δημιουργώ τον πίνακα που θα περιλαμβάνει τις τιμές των
εργαστηριακών μετρήσεων των ασθενών με εγκεφαλικό που
έχουν εισαχθεί στις ΜΕΘ
"""

import pandas as pd

#Δημιουργώ τις στήλες από το αρχείο που περιλαμβάνει τις μοναδικές εξετάσεις
df_columns = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\o12_cerebral_infarction_labevents_counts.csv',
                        usecols = ['itemid'], index_col=False)

#Ορίζω το header ως index έτσι ώστε όταν αλλάξω την κατεύθυνση του csv να μπουν τα στοιχεία στην κεφαλίδα.
df_columns = df_columns.set_index('itemid')

# Μεταφέρω τις γραμμές σε στήλες
df_columns = df_columns.transpose()

df_rows = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\o11_cerebral_infarction_labevents_grouped_8hours.csv')

# Μετατροπή της στήλης charttime από object σε datetime για να μπορέσει να γίνει ο παρακάτω διαχωρισμός σε δυο στήλες
df_rows['charttime']= pd.to_datetime(df_rows['charttime'])

#Διαχωρισμός σε δυο καινούργιες στήλες της ημερομηνίας και ώρας
df_rows['date'] = pd.to_datetime(df_rows['charttime']).dt.date
df_rows['time'] = pd.to_datetime(df_rows['charttime']).dt.time

#Αλλαγή της σειράς εμφάνισης των στοιχείων του dataframe,
#για να είναι ευανάγνωστο κατά τους ελέγχους μου,
#δεν είναι απαραίτητο για την εξέλιξη του κώδικα
df_rows = df_rows[['subject_id', 'hadm_id', 'charttime', 'date', 'time', 'itemid', 'Mean_Lab']]

# Αποθηκεύω το αρχείο που πλεον περιλαμβάνει τις δυο καινούργιες στήλες,
# date και time
df_rows.to_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\o13_cerebral_infarction_labevents_grouped_8hours_date_time.csv', index=False)

# Εδώ παίρνουμε τις τιμές της στήλης time και βλέπουμε πως γκρουπάρονται οι ώρες.
# Οπ΄ότε έχουμε τρία οχτάωρα 00:00, 08:00 και 16:00
list(set(df_rows['time']))

# Απλά προβάλω τις στήλες με τα στοιχεία των εξετάσεων
df_columns

"""
Δημιουργώ ένα μοναδικό αναγνωριστικό με την ένωση των
στοιχείων από στήλες subject, hadm_id, date και time.
Ανάμεσα στα στοιχεία προβάλω ένα κενό το οποίο θα μου
φανεί χρήσιμο αργότερα για τον διαχωρισμό τους ξανά σε
διαφορετικές στήλες.
"""
df_rows["concat"] = (df_rows["subject_id"].astype(str)
                     +" "
                     +df_rows["hadm_id"].astype(str)
                     +" "
                     +df_rows["date"].astype(str)
                     +" "+
                     df_rows["time"].astype(str))

# Ελέγχω αν υπάρχει κενό κελί στην στήλη concat.
df_rows['concat'].isnull().values.any()

# Αλλάζω την σειρά των στηλών του dataframe
df_rows = df_rows[['concat', 'subject_id', 'hadm_id', 'charttime', 'date', 'time', 'itemid', 'Mean_Lab']]

# Από το dr_rows κρατάω τις μοναδικές εγγραφές
df_copy_rows = pd.DataFrame(df_rows.concat.unique())

# Δημιουργώ καινούργιο πίνακα ενσωματώνοντας στον πίνακα df columns τις στήλες με τις κενές για την ώρα εξετάσεις
lab_tbl = df_columns.append(df_copy_rows)

# Μετονομάζω την στήλη 0 σε concat για να είναι ίδιες στους δυο πίνακες
lab_tbl = lab_tbl.rename({0:'concat'}, axis=1)

#Διαχωρισμός του concat σε τέσσερις στήλες
lab_tbl[['subject_id','hadm_id', 'date', 'time']] = lab_tbl['concat'].str.split(expand=True)

#Διαγράφω την στήλη concat γιατί θα πρέπει να την ξαναδημιουργήσω
#χωρίς κενά ανάμεσα στα ενοποιημένα στοιχεία, οπότε θα την έχω διπλή.
lab_tbl = lab_tbl.drop('concat', 1)

# Τρέχω το αποτέλεσμα του πίνακα
lab_tbl

lab_tbl.to_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\o14_empty_lab_tbl.csv', index=False)
