==> Old_Code/03_05_lab_table_label_header.ipynb <==
"""
Αντικατάσταση των κεφαλίδων στο αρχείο
με τις συνολικές καταγραφές
"""

#Εισαγωγή βιβλιοθήκης Panda
import pandas as pd

df_data_labevents = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\hosp\d_labitems.csv.gz')
df_item_header = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\o15_fill_lab_itemid_header_tbl.csv')

df_data_labevents = df_data_labevents.set_index('itemid')

# Ενοποίηση τιμών σε κελιά
df_data_labevents["header"] = (df_data_labevents["label"].astype(str)
                                                + " ("
                                                + df_data_labevents["fluid"].astype(str)
                                                + ")")

df_data_labevents

# Ελέγχω τις κεφαλίδες και αν δεν ανήκουν στις
# τέσσερις πρώτες τιμές κάνω τις αντικατάσταση.
for col in df_item_header.columns:
    if ((col != 'subject_id')
        and (col != 'hadm_id')
        and (col != 'Department')
        and (col != 'date')
        and (col != 'time')):
        temp = (df_data_labevents.loc[int(col), 'header'])
        df_item_header.rename(columns={'{}'.format(col): '{}'.format(temp)}, inplace=True)

df_item_header

# Αποθήκευση του τελικού αρχείου με επικεφαλίδες τους κωδικούς των παρατηρήσεων
df_item_header.to_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\o15_fill_lab_label_header_tbl.csv', index=False)
