==> Old_Code/04_01_ED_patients.ipynb <==
"""
Καταγράφω τους ασθενείς που πέρασαν από τα
επείγοντα με εγκεφαλικό και κατέληξαν στις ΜΕΘ.


2.292 -> Ασθενείς με εγκεφαλικό στα επείγοντα.
1.547 -> Συνολικές εισαγωγές στις ΜΕΘ από τα επείγοντα.
1.546 -> Μοναδικές εισαγωγές στις ΜΕΘ από τα επείγοντα.
1.500 -> Μοναδικών ασθενών στις ΜΕΘ.


"""

import pandas as pd

df_patients = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\ed\diagnosis.csv.gz')
df_icu_patients = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\o5_icu_unique_admissions.csv')
df_edstays = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\ed\edstays.csv.gz')

#Ασθενείς με εγκεφαλικό.
ed_cerebral_infarction_diagnoses = df_patients[
    (df_patients['icd_code'] =="43300") | (df_patients['icd_code'] =="43301") | (df_patients['icd_code'] =="43310") | (df_patients['icd_code'] =="43311") |
    (df_patients['icd_code'] =="43320") | (df_patients['icd_code'] =="43321") | (df_patients['icd_code'] =="43330") | (df_patients['icd_code'] =="43331") |
    (df_patients['icd_code'] =="43380") | (df_patients['icd_code'] =="43381") | (df_patients['icd_code'] =="43390") | (df_patients['icd_code'] =="43391") |
    (df_patients['icd_code'] =="43400") | (df_patients['icd_code'] =="43401") | (df_patients['icd_code'] =="43410") | (df_patients['icd_code'] =="43411") |
    (df_patients['icd_code'] =="43491") | (df_patients['icd_code'] =="43490") | (df_patients['icd_code'] =="436") | (df_patients['icd_code'] =="431") |
    (df_patients['icd_code'] =="430") | (df_patients['icd_code'] =="4350") | (df_patients['icd_code'] =="4351") | (df_patients['icd_code'] =="4352") |
    (df_patients['icd_code'] =="4353") | (df_patients['icd_code'] =="4358") | (df_patients['icd_code'] =="4359")
]

# Σύνολο εισαγωγών στα επείγοντα με εγκεφαλικό
ed_cerebral_infarction_diagnoses.info()

rslt_df=pd.merge(df_icu_patients, df_edstays, on=['subject_id', 'hadm_id'])

# Συνολικές εισαγωγές στις ΜΕΘ από τα επείγοντα
rslt_df.info()

#Μοναδικοί ασθενείς στις ΜΕΘ από τα επείγοντα
rslt_df.drop_duplicates(subset=['subject_id']).info()

#Εξαγωγή σε csv
rslt_df.to_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\o16_ed_patients_to_icu.csv', index=False)
