==> Old_Code/04_02_ED_mean_table.ipynb <==
"""
1.651.119 -> Στο σύνολο γραμμές καταχωρημένων ενδείξεων.
752.146 -> Ομαδοποιημένες εγγραφές στο 8ώρο.
1.547 -> Eισαγωγές στις ΜΕΘ
1.662 -> Ομαδοποιημένες εγγραφές στο 8ώρο που αφορούν ασθενείς από τα επείγοντα στις ΜΕΘ
"""

import pandas as pd

df_vitalsign = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\ed\Vitalsign.csv.gz')
df_ed_to_icu = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\o16_ed_patients_to_icu.csv', usecols = ['subject_id', 'hadm_id','stay_id'])

# Μετατροπή της στήλης charttime από object σε datetime
df_vitalsign['charttime']= pd.to_datetime(df_vitalsign['charttime'])

# Ομαδοποίηση καταγραφών ανά στήλη στο 8ώρο.

grouped_multiple = (df_vitalsign.groupby(['subject_id', 'stay_id',
                                          pd.Grouper(key='charttime',
                                                     freq='8H')]).agg({'temperature': ['mean'],
                                                                       'heartrate': ['mean'],
                                                                       'resprate': ['mean'],
                                                                       'o2sat': ['mean'],
                                                                       'sbp': ['mean'],
                                                                       'dbp': ['mean']}))
grouped_multiple.columns = (['Mean_temperature',
                             'Mean_heartrate',
                             'Mean_resprate',
                             'Mean_o2sat',
                             'Mean_sbp',
                             'Mean_dbp'])
grouped_multiple = grouped_multiple.reset_index()

# Κενές τιμές στον πίνακα
grouped_multiple.isna().sum()

df_ed_to_icu.info()

rslt_df=pd.merge(df_ed_to_icu, grouped_multiple, on=['stay_id'])

#Διαχωρισμός σε δυο καινούργιες στήλες της ημερομηνίας και ώρας
rslt_df['date'] = pd.to_datetime(rslt_df['charttime']).dt.date
rslt_df['time'] = pd.to_datetime(rslt_df['charttime']).dt.time

rslt_df = rslt_df.drop(['subject_id_y', 'stay_id', 'charttime'], axis=1)

rslt_df = rslt_df.rename(columns={"subject_id_x": "subject_id"})

rslt_df.insert(2, 'Department', 'ed')

rslt_df

#Εξαγωγή σε csv
rslt_df.to_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\o17_ed_to_icu_vitalsigns.csv', index=False)
