==> Old_Code/05_01_Final_Table.ipynb <==
"""
Ενοποίηση και των τριών πινάκων που έχω δημιουργήσει σε ένα ενιαίο

ChartEvent -> 65.189 γραμμές / 712 στήλες
LabEvent -> 68.191 γραμμές / 393 στήλες
ED vitals -> 1.662 γραμμές / 11 στήλες

final_table -> 135.042 γραμμές / 1106 στήλες
"""

import pandas as pd

df_chartevent = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\o9_fill_chart_label_header_tbl.csv')

df_labevents = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\o15_fill_lab_label_header_tbl.csv')

df_vital = pd.read_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\o17_ed_to_icu_vitalsigns.csv')

final_table = df_chartevent.append([df_labevents, df_vital])

#Εξαγωγή σε csv
final_table.to_csv('D:\MSc Aegean\Διατριβή\MIMIC_IV\mimic-iv-1.0\exports\o18_final_table.csv', index=False)
